+++ src/main/java/cart/controller/MainController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
-import cart.dto.ProductsResponseDto;
+import cart.dto.product.ProductsResponseDto;
 import cart.service.ProductService;
 
 @Controller
