+++ src/main/java/cart/dto/product/ProductCreateRequestDto.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 import cart.domain.Product;
