+++ src/main/java/cart/dto/product/ProductUpdateRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 public class ProductUpdateRequest {
 	private final String name;
