+++ src/main/java/cart/dto/product/ProductsResponseDto.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 import java.util.List;
 import java.util.stream.Collectors;
